+++ src/interrupt.rs
@@ -1,9 +1,6 @@
 //! Various functions related to interrupts
 //!
-//! TO DO: 
-//! - "enable" flag in the interrupt enable function must be moved to the config, called "enable_differential"
-//! - interrupt status reading must be in one go, as with latching on it would clear the register by reading the first bit the way it's configured now
-//! 
+//! TO DO: check if all functions related to the interrupts are covered
 
 use super::*;
 
@@ -81,6 +78,9 @@ impl InterruptConfig {
         
         let mut data = 0u8;
 
+        if self.enable_differential.status() {
+            data |=1 << 3;
+        }
         if self.enable_latch_interrupt.status() {
             data |= 1 << 2;
         }
@@ -107,100 +107,30 @@ where
     T: Interface<Error = E>,
 {
     /// Enable interrupts and configure the interrupt pin
-    pub fn configure_interrupts(&mut self, flag: bool, config: InterruptConfig,) -> Result<(), T::Error> {
-        match config.enable_differential {
-            FLAG::Enabled => self.set_register_bit_flag(Registers::INTERRUPT_CFG, Bitmasks::DIFF_EN),
-            FLAG::Disabled => self.clear_register_bit_flag(Registers::INTERRUPT_CFG, Bitmasks::DIFF_EN),
-        }?;
-        self.interface
-            .write(Registers::CTRL_REG3.addr(), config.int_ctrl_reg3())?;        
+    pub fn configure_interrupts(&mut self, 
+                                // flag: bool, 
+                                config: InterruptConfig)
+                                 -> Result<(), T::Error> {        
         
+        // write the whole CTRL_REG3 register                                    
+        self.interface.write(Registers::CTRL_REG3.addr(), config.int_ctrl_reg3())?;                
+        
+        // get the contents of INTERRUPT_CFG and combine it with the bits to be set
         let mut buffer = [0u8;1];
         self.read_register(Registers::INTERRUPT_CFG)?;        
-        let mut interrupt_cfg = 0u8;
-        interrupt_cfg |= config.int_interrupt_cfg();
-       
-        self.interface
-            .write(Registers::INTERRUPT_CFG.addr(), interrupt_cfg)?;
+                
+        let mut interrupt_cfg = config.int_interrupt_cfg();
+               
+        buffer = buffer & !0b00001111;
+
+        buffer = buffer | interrupt_cfg;
+
+        self.interface.write(Registers::INTERRUPT_CFG.addr(), buffer)?;        
+        
         Ok(())
     }
     
-    // --- THE FOLLOWING SECTION COULD BE REMOVED --- 
-
-    /*
     
-    
-    /// Configuration of the interrupt generation (enabled/disable)
-    pub fn int_generation_enable(&mut self, flag: bool) -> Result<(), T::Error> {
-        match flag {
-            true => self.set_register_bit_flag(Registers::INTERRUPT_CFG, Bitmasks::DIFF_EN),
-            false => self.clear_register_bit_flag(Registers::INTERRUPT_CFG, Bitmasks::DIFF_EN),
-        }
-    }
-
-    /// Interrupt request latching to INT_SOURCE
-    pub fn int_latch_enable(&mut self, flag: bool) -> Result<(), T::Error> {
-        match flag {
-            true => self.set_register_bit_flag(Registers::INTERRUPT_CFG, Bitmasks::LIR),
-            false => self.clear_register_bit_flag(Registers::INTERRUPT_CFG, Bitmasks::LIR),
-        }
-    }
-
-    /// Data-ready signal on INT_DRDY pin
-    pub fn data_signal_drdy_enable(&mut self, flag: bool) -> Result<(), T::Error> {
-        match flag {
-            true => self.set_register_bit_flag(Registers::CTRL_REG3, Bitmasks::DRDY),
-            false => self.clear_register_bit_flag(Registers::CTRL_REG3, Bitmasks::DRDY),
-        }
-    }
-
-    /// Interrupt active high/low (default active high)
-    pub fn interrupt_pin_active(&mut self, setting: INT_ACTIVE) -> Result<(), T::Error> {
-        match setting {
-            INT_ACTIVE::High => {
-                self.clear_register_bit_flag(Registers::CTRL_REG3, Bitmasks::INT_H_L)
-            }
-            INT_ACTIVE::Low => self.set_register_bit_flag(Registers::CTRL_REG3, Bitmasks::INT_H_L),
-        }
-    }
-
-    /// Interrupt pin configuration: push-pull (default) or open drain
-    pub fn interrupt_pin_config(&mut self, setting: INT_PIN) -> Result<(), T::Error> {
-        match setting {
-            INT_PIN::PushPull => {
-                self.clear_register_bit_flag(Registers::CTRL_REG3, Bitmasks::PP_OD)
-            }
-            INT_PIN::OpenDrain => self.set_register_bit_flag(Registers::CTRL_REG3, Bitmasks::PP_OD),
-        }
-    }
-
-    /// Configure INT_DRDY pin
-    pub fn int_drdy_config(&mut self, config: INT_DRDY) -> Result<(), T::Error> {
-        let mut reg_data = [0u8];
-        self.interface
-            .read(Registers::CTRL_REG3.addr(), &mut reg_data)?;
-        let mut payload = reg_data[0];
-        payload &= !Bitmasks::INT_S_MASK;
-        payload |= config.value();
-        self.interface.write(Registers::CTRL_REG3.addr(), payload)?;
-        Ok(())
-    }
-
-    */
-    /*
-    /// Get all the flags from the INT_SOURCE register (NOTE: INT_SOURCE register is cleared by reading it)
-    pub fn get_int_status(&mut self) -> Result<IntSource, T::Error> {        
-        let status = IntSource {
-            /// Has any interrupt event been generated? (self clearing)
-            interrupt_active: self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::IA)?,
-            /// Has low differential pressure event been generated? (self clearing)
-            diff_press_low: self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::PL)?,
-            /// Has high differential pressure event been generated? (self clearing)
-            diff_press_high: self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::PH)?,
-        };
-        Ok(status)
-    }
- */
  /// Get all the flags from the INT_SOURCE register (NOTE: INT_SOURCE register is cleared by reading it)
  pub fn get_int_status(&mut self) -> Result<IntStatus, T::Error> {        
                 
@@ -227,31 +157,4 @@ where
  }
 
 }
-    // --- THESE FUNCTIONS COULD BE REMOVED ---
-
-    /*
-
-    /// Has any interrupt event been generated? (self clearing)
-    pub fn interrupt_active(&mut self) -> Result<bool, T::Error> {
-        self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::IA)
-    }
-
-    /// Has low differential pressure event been generated? (self clearing)
-    pub fn low_pressure_event_occurred(&mut self) -> Result<bool, T::Error> {
-        self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::PL)
-    }
-
-    /// Has high differential pressure event been generated? (self clearing)
-    pub fn high_pressure_event_occurred(&mut self) -> Result<bool, T::Error> {
-        self.is_register_bit_flag_high(Registers::INT_SOURCE, Bitmasks::PH)
-    }
-
-    /// Enable interrupt on differential pressure high event
-    pub fn diff_press_high_enable(&mut self, flag: bool) -> Result<(), T::Error> {
-        match flag {
-            true => self.set_register_bit_flag(Registers::INTERRUPT_CFG, Bitmasks::PHE),
-            false => self.clear_register_bit_flag(Registers::INTERRUPT_CFG, Bitmasks::PHE),
-        }
-    }
-
-     */
+