+++ src/sensor.rs
@@ -25,9 +25,6 @@ where
         Ok(whoami)
     }
 
-
-    /* */
-
      /// Calculated pressure reading in hPa
      pub fn read_pressure(&mut self) -> Result<f32, T::Error> {
         let mut data = [0u8; 3];
@@ -51,34 +48,7 @@ where
         let temperature = (t as f32) / TEMP_SCALE;
         Ok(temperature)
     }
-
-    /*
-    /// Raw sensor reading (3 bytes of pressure data and 2 bytes of temperature data)
-    fn read_sensor_raw(&mut self) -> Result<(i32, i32), T::Error> {
-        let mut data = [0u8; 5];
-        self.interface
-            .read(Registers::PRESS_OUT_XL.addr(), &mut data)?;
-        let p: i32 = (data[2] as i32) << 16 | (data[1] as i32) << 8 | (data[0] as i32);
-        let t: i32 = (data[4] as i32) << 8 | (data[3] as i32);
-        Ok((p, t))
-    }
-
-    /// Calculated pressure reading in hPa
-    pub fn read_pressure(&mut self) -> Result<f32, T::Error> {
-        let (p, _t) = self.read_sensor_raw()?;
-        let pressure: f32 = (p as f32) / PRESS_SCALE;
-        Ok(pressure)
-    }
-
-    /// Calculated temperaure reading in degrees Celsius
-    pub fn read_temperature(&mut self) -> Result<f32, T::Error> {
-        let (_p, t) = self.read_sensor_raw()?;
-        let temperature: f32 = (t as f32) / TEMP_SCALE;
-        Ok(temperature)
-    }
-
-     */
-
+  
     /// Calculated reference pressure reading in hPa
     pub fn read_reference_pressure(&mut self) -> Result<f32, T::Error> {
         let mut data = [0u8; 3];
@@ -132,24 +102,32 @@ where
         Ok(())
     }
 
-    /*
+    /// Set the reference pressure (value in hPA)
+    pub fn set_reference_pressure(&mut self, pressure: u16) -> Result<(), T::Error> {
+        
+        /*
+        self.interface.read(Registers::REF_P_XL.addr(), &mut data)?;
+        let p: i32 = (data[2] as i32) << 16 | (data[1] as i32) << 8 | (data[0] as i32);
+        let pressure: f32 = (p as f32) / PRESS_SCALE;
+        
+        
+        let pressure = pressure * PRESS_SCALE;
 
-    /// Get all the flags from the STATUS_REG register
-    pub fn get_data_status(&mut self) -> Result<DataStatus, T::Error> {
-        let status = DataStatus {
-            /// Has new temperature data overwritten the previous one?
-            temp_overrun: self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::T_OR)?,
-            /// Has new pressure data overwritten the previous one?
-            press_overrun: self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::P_OR)?,
-            /// Is new temperature data available?
-            temp_available: self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::T_DA)?,
-            /// Is new pressure data available?            
-            press_available: self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::P_DA)?,
-        };
-        Ok(status)
+        let mut payload = [0u8; 3];
+        
+        // value must be split into three bytes
+
+        payload[0] = (pressure & 0xff) as u8; // XL byte
+        payload[1] = (pressure >> 8) as u8; // L byte
+        payload[2] = (pressure >> 16) as u8; // H byte
+                
+        */
+        
+
+
+        
+        Ok(())
     }
-
-     */
 
     /// Get all the flags from the STATUS_REG register
     pub fn get_data_status(&mut self) -> Result<DataStatus, T::Error> {
@@ -193,29 +171,5 @@ where
         Ok(())
     }
 
-    // --- THESE FUNCTIONS CAN BE REMOVED ---
 
-    /*
-
-    /// Has new pressure data overwritten the previous one?
-    pub fn pressure_data_overrun(&mut self) -> Result<bool, T::Error> {
-        self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::P_OR)
-    }
-
-    /// Has new temperature data overwritten the previous one?
-    pub fn temperature_data_overrun(&mut self) -> Result<bool, T::Error> {
-        self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::T_OR)
-    }
-
-    /// Is new pressure data available?
-    pub fn pressure_data_available(&mut self) -> Result<bool, T::Error> {
-        self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::P_DA)
-    }
-
-    /// Is new temperature data available?
-    pub fn temperature_data_available(&mut self) -> Result<bool, T::Error> {
-        self.is_register_bit_flag_high(Registers::STATUS, Bitmasks::T_DA)
-    }
-
-     */
 }